+++ modules/github-iam-oidc/github.repo.role.tf
@@ -1,6 +1,6 @@
 resource "aws_iam_role" "this" {
-  for_each    = { for index, github_repo in var.github_repositories : github_repo.name => github_repo }
-  name        = module.iamroles_label[each.value.name].id
+  for_each    = var.github_repositories
+  name        = "${var.name_prefix}-${each.key}"
   description = "Role for Github Actions"
   assume_role_policy = jsonencode({
     Version = "2012-10-17"
@@ -11,7 +11,7 @@ resource "aws_iam_role" "this" {
         Effect = "Allow"
         Condition = {
           StringLike = {
-            "token.actions.githubusercontent.com:sub" = "repo:${var.github_org}/${each.value.name}:*"
+            "token.actions.githubusercontent.com:sub" = "repo:${var.github_org}/${each.key}:*"
           }
           StringEquals = {
             "token.actions.githubusercontent.com:aud" = "sts.amazonaws.com"
@@ -23,33 +23,16 @@ resource "aws_iam_role" "this" {
       },
     ]
   })
-  tags       = module.iamroles_label[each.value.name].tags
   path       = "/github/"
   depends_on = [aws_iam_openid_connect_provider.this]
 }
 
 resource "aws_iam_role_policy_attachment" "this" {
-  for_each = tomap({
-    for policy in local.policies : "${policy.repo}.${policy.policy_arn}" => policy
-  })
-  policy_arn = each.value.policy_arn
-  role       = module.iamroles_label[each.value.repo].id
-  depends_on = [aws_iam_role.this]
-}
-
-module "iamroles_label" {
-  source  = "cloudposse/label/null"
-  version = "~> 0.24.1"
-
-  for_each = { for index, github_repo in var.github_repositories : github_repo.name => github_repo }
-
-  namespace = "gs"
-  stage     = "prod"
-  name      = each.value.name
-  delimiter = "-"
-
-  tags = {
-    terraform   = "true",
-    environment = var.environment
+  for_each = {
+    for idx, item in local.github_policy_attachments :
+    "${item.repo_name}-${idx}" => item
   }
+  policy_arn = each.value.policy_arn
+  role       = aws_iam_role.github[each.value.repo_name].name
+  depends_on = [aws_iam_role.this]
 }