+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
   - Before upgrading, run: `terraform state mv aws_key_pair.this aws_key_pair.this[0]` to update tfstate
 - Made `module.ecs-capacity-provider.key_pair_name` optional.
 - Made force_delete on the ASG for `module.ecs-capacity-provider`.
+- Fix on `module.ecr` removed hardcoded prefix, using `name_prefix` instead.
+  - Breaking change. Pending to test how to prevent issues.
 
 ### Security
 - (any security fixes)